Reorganized docs and cleaned files
