+++ server/src/app.js
@@ -10,6 +10,7 @@ require('dotenv').config();
 const middlewares = require('./middlewares');
 const api = require('./api');
 const routes = require('./routes');
+const swaggerOptions = require('./docs/swagger');
 
 const app = express();
 
@@ -18,59 +19,11 @@ app.use(helmet());
 app.use(cors());
 app.use(express.json());
 
-
-//import schema
-const Posts = require('../models/Posts');
-const myObj = {
-  "type": "object",
-  "properties": {
-     "title": {
-        "type": "string",
-        "example": "The best title"
-     },
-     "content": {
-        "type": "string",
-        "example": "The best content"
-     }
-  }
-};
-
-// Extended: https://swagger.io/specification/#infoObject
-const swaggerOptions = {
-  swaggerDefinition: {
-    info: {
-      title: 'RumbaPR API',
-      description: 'The REST API for the RumbaPR web application.',
-      version: '1.0.0',
-      contact: {
-        name: 'The White Lotus'
-      },
-    },
-    basePath: '/api/v2',
-    servers: [
-      { 
-        url: "http://localhost:5000/api/v2",
-        description: "Local server"
-      }
-    ],
-    openapi: "3.0.0",
-    "components": {
-      "schemas": {
-         "Posts": {...myObj} 
-      }
-    }
-  },
-  // Directory to routes
-  // ex to find within a folder ['./routes/*.js']
-  apis: ["./src/api/posts.js", "../models/Posts"]
-};
-
 const swaggerDocs = swaggerJsDoc(swaggerOptions);
 app.use('/api-docs', swaggerUI.serve, swaggerUI.setup(swaggerDocs));
 
 
 app.get('/', (req, res) => {
-  console.log(Posts.schema.obj);
   res.json({
     message: '✨🌎Server is running!🌏✨'
   });
